Refactor primitive utilities to a shared module
