+++ legend-pure-core/legend-pure-m3-core/src/main/resources/platform/pure/grammar/functions/meta/type/relation/window/frame.pure
@@ -29,6 +29,11 @@ Class meta::pure::functions::relation::FrameIntValue extends meta::pure::functio
     value : Integer[1];
 }
 
+Class meta::pure::functions::relation::FrameNumericValue extends meta::pure::functions::relation::FrameValue
+{
+    value : Number[1];
+}
+
 Class meta::pure::functions::relation::UnboundedFrameValue extends meta::pure::functions::relation::FrameValue
 {
 }
