+++ legend-pure-runtime/legend-pure-runtime-java-engine-interpreted/src/main/java/org/finos/legend/pure/runtime/java/interpreted/natives/NumericUtilities.java
@@ -15,19 +15,11 @@
 package org.finos.legend.pure.runtime.java.interpreted.natives;
 
 import org.eclipse.collections.api.block.predicate.Predicate;
-import org.eclipse.collections.api.factory.Lists;
-import org.eclipse.collections.api.list.ListIterable;
-import org.finos.legend.pure.m3.navigation.M3Paths;
-import org.finos.legend.pure.m3.navigation.M3Properties;
 import org.finos.legend.pure.m3.navigation.PrimitiveUtilities;
 import org.finos.legend.pure.m3.navigation.ProcessorSupport;
 import org.finos.legend.pure.m3.navigation.ValueSpecificationBootstrap;
 import org.finos.legend.pure.m4.ModelRepository;
 import org.finos.legend.pure.m4.coreinstance.CoreInstance;
-import org.finos.legend.pure.m4.coreinstance.primitive.DecimalCoreInstance;
-import org.finos.legend.pure.m4.coreinstance.primitive.FloatCoreInstance;
-import org.finos.legend.pure.m4.coreinstance.primitive.IntegerCoreInstance;
-import org.finos.legend.pure.m4.coreinstance.primitive.PrimitiveCoreInstance;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
@@ -40,13 +32,13 @@ public class NumericUtilities
     @Deprecated
     public static Predicate<CoreInstance> IS_DECIMAL_CORE_INSTANCE(ProcessorSupport processorSupport)
     {
-        return instance -> isDecimal(instance, processorSupport);
+        return instance -> PrimitiveUtilities.isDecimal(instance, processorSupport);
     }
 
     @Deprecated
     public static Predicate<CoreInstance> IS_FLOAT_CORE_INSTANCE(ProcessorSupport processorSupport)
     {
-        return instance -> isFloat(instance, processorSupport);
+        return instance -> PrimitiveUtilities.isFloat(instance, processorSupport);
     }
 
     private NumericUtilities()
@@ -202,184 +194,5 @@ public class NumericUtilities
     public static CoreInstance newPureInteger(int number, ModelRepository repository)
     {
         return repository.newIntegerCoreInstance(number);
-    }
-
-    /**
-     * Convert a Pure number into a Java number.
-     *
-     * @param pureNumber       Pure number
-     * @param processorSupport processor support
-     * @return Java number
-     */
-    public static Number toJavaNumber(CoreInstance pureNumber, ProcessorSupport processorSupport)
-    {
-        CoreInstance classifier = processorSupport.getClassifier(pureNumber);
-        if (classifier.getValueForMetaPropertyToOne(M3Properties._package) == null)
-        {
-            String typeName = classifier.getName();
-            if (typeName != null)
-            {
-                switch (typeName)
-                {
-                    case M3Paths.Integer:
-                    {
-                        return PrimitiveUtilities.getIntegerValue(pureNumber);
-                    }
-                    case M3Paths.Float:
-                    {
-                        return PrimitiveUtilities.getFloatValue(pureNumber);
-                    }
-                    case M3Paths.Decimal:
-                    {
-                        return PrimitiveUtilities.getDecimalValue(pureNumber);
-                    }
-                }
-            }
-        }
-        else
-        {
-            if (isInteger(pureNumber, processorSupport))
-            {
-                return PrimitiveUtilities.getIntegerValue(pureNumber);
-            }
-            else if (isFloat(pureNumber, processorSupport))
-            {
-                return PrimitiveUtilities.getFloatValue(pureNumber);
-            }
-            else if (isDecimal(pureNumber, processorSupport))
-            {
-                return PrimitiveUtilities.getDecimalValue(pureNumber);
-            }
-        }
-        throw new IllegalArgumentException("Not a number: " + pureNumber);
-    }
-
-    public static boolean isInteger(CoreInstance instance, ProcessorSupport processorSupport)
-    {
-        return isNumberOfType(instance, IntegerCoreInstance.class, M3Paths.Integer, processorSupport);
-    }
-
-    public static boolean isFloat(CoreInstance instance, ProcessorSupport processorSupport)
-    {
-        return isNumberOfType(instance, FloatCoreInstance.class, M3Paths.Float, processorSupport);
-    }
-
-    public static boolean isDecimal(CoreInstance instance, ProcessorSupport processorSupport)
-    {
-        return isNumberOfType(instance, DecimalCoreInstance.class, M3Paths.Decimal, processorSupport);
-    }
-
-    private static boolean isNumberOfType(CoreInstance instance, Class<? extends PrimitiveCoreInstance<? extends Number>> coreInstanceClass, String numberTypeName, ProcessorSupport processorSupport)
-    {
-        if (coreInstanceClass.isInstance(instance))
-        {
-            return true;
-        }
-        CoreInstance classifier = processorSupport.getClassifier(instance);
-        return (classifier.getValueForMetaPropertyToOne(M3Properties._package) == null) && numberTypeName.equals(classifier.getName());
-    }
-
-    /**
-     * Convert a collection of Pure numbers into a collection of Java numbers.
-     *
-     * @param pureNumbers Pure numbers
-     * @return Java numbers
-     */
-    public static ListIterable<Number> toJavaNumber(ListIterable<? extends CoreInstance> pureNumbers, ProcessorSupport processorSupport)
-    {
-        return pureNumbers.collect(n -> toJavaNumber(n, processorSupport), Lists.mutable.ofInitialCapacity(pureNumbers.size()));
-    }
-
-    /**
-     * Compare two numbers. Returns a negative value if left is less than right, positive if right is less than left,
-     * or 0 if they are equal.
-     *
-     * @param left  left number
-     * @param right right number
-     * @return comparison
-     */
-    public static int compare(Number left, Number right)
-    {
-        if ((left instanceof Integer) || (left instanceof Long))
-        {
-            if ((right instanceof Integer) || (right instanceof Long))
-            {
-                return Long.compare(left.longValue(), right.longValue());
-            }
-            if ((right instanceof Float) || (right instanceof Double))
-            {
-                return BigDecimal.valueOf(left.longValue()).compareTo(BigDecimal.valueOf(right.doubleValue()));
-            }
-            if (right instanceof BigInteger)
-            {
-                return BigInteger.valueOf(left.longValue()).compareTo((BigInteger) right);
-            }
-            if (right instanceof BigDecimal)
-            {
-                return BigDecimal.valueOf(left.longValue()).compareTo((BigDecimal) right);
-            }
-            throw new RuntimeException("Number of an unhandled type: " + right);
-        }
-        if ((left instanceof Float) || (left instanceof Double))
-        {
-            if ((right instanceof Integer) || (right instanceof Long))
-            {
-                return BigDecimal.valueOf(left.doubleValue()).compareTo(BigDecimal.valueOf(right.longValue()));
-            }
-            if ((right instanceof Float) || (right instanceof Double))
-            {
-                return Double.compare(left.doubleValue(), right.doubleValue());
-            }
-            if (right instanceof BigInteger)
-            {
-                return BigDecimal.valueOf(left.doubleValue()).compareTo(new BigDecimal((BigInteger) right)); //NOSONAR
-            }
-            if (right instanceof BigDecimal)
-            {
-                return BigDecimal.valueOf(left.doubleValue()).compareTo((BigDecimal) right);
-            }
-            throw new RuntimeException("Number of an unhandled type: " + right);
-        }
-        if (left instanceof BigInteger)
-        {
-            if ((right instanceof Integer) || (right instanceof Long))
-            {
-                return ((BigInteger) left).compareTo(BigInteger.valueOf(right.longValue()));
-            }
-            if ((right instanceof Float) || (right instanceof Double))
-            {
-                return new BigDecimal((BigInteger) left).compareTo(BigDecimal.valueOf(right.doubleValue()));
-            }
-            if (right instanceof BigInteger)
-            {
-                return ((BigInteger) left).compareTo((BigInteger) right);
-            }
-            if (right instanceof BigDecimal)
-            {
-                return new BigDecimal((BigInteger) left).compareTo((BigDecimal) right);
-            }
-            throw new RuntimeException("Number of an unhandled type: " + right);
-        }
-        if (left instanceof BigDecimal)
-        {
-            if ((right instanceof Integer) || (right instanceof Long))
-            {
-                return ((BigDecimal) left).compareTo(BigDecimal.valueOf(right.longValue()));
-            }
-            if ((right instanceof Float) || (right instanceof Double))
-            {
-                return ((BigDecimal) left).compareTo(BigDecimal.valueOf(right.doubleValue()));
-            }
-            if (right instanceof BigInteger)
-            {
-                return ((BigDecimal) left).compareTo(new BigDecimal((BigInteger) right));
-            }
-            if (right instanceof BigDecimal)
-            {
-                return ((BigDecimal) left).compareTo((BigDecimal) right);
-            }
-            throw new RuntimeException("Number of an unhandled type: " + right);
-        }
-        throw new RuntimeException("Number of an unhandled type: " + left);
     }
 }
