+++ legend-pure-runtime/legend-pure-runtime-java-engine-interpreted/src/main/java/org/finos/legend/pure/runtime/java/interpreted/natives/grammar/math/operation/Minus.java
@@ -20,6 +20,7 @@ import org.eclipse.collections.api.stack.MutableStack;
 import org.finos.legend.pure.m3.compiler.Context;
 import org.finos.legend.pure.m3.navigation.Instance;
 import org.finos.legend.pure.m3.navigation.M3Properties;
+import org.finos.legend.pure.m3.navigation.PrimitiveUtilities;
 import org.finos.legend.pure.m3.navigation.ProcessorSupport;
 import org.finos.legend.pure.m3.navigation.ValueSpecificationBootstrap;
 import org.finos.legend.pure.m4.ModelRepository;
@@ -57,16 +58,16 @@ public class Minus extends NativeFunction
             {
                 CoreInstance number = numbers.get(0);
                 NumericAccumulator accumulator = NumericAccumulator.newAccumulator(0);
-                accumulator.subtract(NumericUtilities.toJavaNumber(number, processorSupport));
-                return NumericUtilities.toPureNumberValueExpression(accumulator.getValue(), NumericUtilities.isDecimal(number, processorSupport), this.repository, processorSupport);
+                accumulator.subtract(PrimitiveUtilities.toJavaNumber(number, processorSupport));
+                return NumericUtilities.toPureNumberValueExpression(accumulator.getValue(), PrimitiveUtilities.isDecimal(number, processorSupport), this.repository, processorSupport);
             }
             default:
             {
-                NumericAccumulator accumulator = NumericAccumulator.newAccumulator(NumericUtilities.toJavaNumber(numbers.get(0), processorSupport));
-                boolean bigDecimalToPureDecimal = numbers.anySatisfy(n -> NumericUtilities.isDecimal(n, processorSupport));
+                NumericAccumulator accumulator = NumericAccumulator.newAccumulator(PrimitiveUtilities.toJavaNumber(numbers.get(0), processorSupport));
+                boolean bigDecimalToPureDecimal = numbers.anySatisfy(n -> PrimitiveUtilities.isDecimal(n, processorSupport));
                 for (int i = 1; i < size; i++)
                 {
-                    accumulator.subtract(NumericUtilities.toJavaNumber(numbers.get(i), processorSupport));
+                    accumulator.subtract(PrimitiveUtilities.toJavaNumber(numbers.get(i), processorSupport));
                 }
                 return NumericUtilities.toPureNumberValueExpression(accumulator.getValue(), bigDecimalToPureDecimal, this.repository, processorSupport);
             }
