+++ legend-pure-runtime/legend-pure-runtime-java-engine-interpreted/src/main/java/org/finos/legend/pure/runtime/java/interpreted/natives/grammar/math/operation/Plus.java
@@ -20,6 +20,7 @@ import org.eclipse.collections.api.stack.MutableStack;
 import org.finos.legend.pure.m3.compiler.Context;
 import org.finos.legend.pure.m3.navigation.Instance;
 import org.finos.legend.pure.m3.navigation.M3Properties;
+import org.finos.legend.pure.m3.navigation.PrimitiveUtilities;
 import org.finos.legend.pure.m3.navigation.ProcessorSupport;
 import org.finos.legend.pure.m3.navigation.ValueSpecificationBootstrap;
 import org.finos.legend.pure.m4.ModelRepository;
@@ -59,11 +60,11 @@ public class Plus extends NativeFunction
             }
             default:
             {
-                NumericAccumulator accumulator = NumericAccumulator.newAccumulator(NumericUtilities.toJavaNumber(numbers.get(0), processorSupport));
-                boolean bigDecimalToPureDecimal = numbers.anySatisfy(n -> NumericUtilities.isDecimal(n, processorSupport));
+                NumericAccumulator accumulator = NumericAccumulator.newAccumulator(PrimitiveUtilities.toJavaNumber(numbers.get(0), processorSupport));
+                boolean bigDecimalToPureDecimal = numbers.anySatisfy(n -> PrimitiveUtilities.isDecimal(n, processorSupport));
                 for (int i = 1; i < size; i++)
                 {
-                    accumulator.add(NumericUtilities.toJavaNumber(numbers.get(i), processorSupport));
+                    accumulator.add(PrimitiveUtilities.toJavaNumber(numbers.get(i), processorSupport));
                 }
                 return NumericUtilities.toPureNumberValueExpression(accumulator.getValue(), bigDecimalToPureDecimal, this.repository, processorSupport);
             }
